Move t helpers into helpers and clean prefixes
