+++ helpers/__init__.py
@@ -6,3 +6,7 @@ from .feature_math import (
     marker_target_aggressive,
     marker_target_conservative,
 )
+from .delete_tracks import delete_selected_tracks
+from .select_short_tracks import select_short_tracks
+from .find_low_marker_frame import find_next_low_marker_frame
+from .set_playhead_to_frame import set_playhead_to_frame
