+++ operators/tracking/export.py
@@ -28,7 +28,7 @@ class CLIP_OT_prefix_new(bpy.types.Operator):
                 _ = track.name
             except Exception as e:
                 print(f"\u26a0\ufe0f Marker-Name fehlerhaft: {track} ({e})")
-                track.name = f"RECOVERED_{i:03d}"
+                track.name = f"{PREFIX_TRACK}{i:03d}"
             else:
                 safe = unicodedata.normalize("NFKD", track.name).encode(
                     "ascii", "ignore"
